AI_FINACE_SYSTEM-main/requirements.txt
google-generativeai
chromadb
langgraph
langchain-google-genai
pandas
numpy
streamlit
plotly
fpdf
python-dotenv
scikit-learn
